requirements.txt
pytest>=8.0
PyYAML>=6.0
requests>=2.31
scikit-learn>=1.5
joblib>=1.4
